Move configure step to runTemplate
