+++ packages/create-cloudflare/src/pages.ts
@@ -1,16 +1,9 @@
-import { chdir } from "process";
 import { crash } from "@cloudflare/cli";
 import { brandColor, dim } from "@cloudflare/cli/colors";
-import {
-	installWrangler,
-	resetPackageManager,
-	retry,
-	runCommand,
-} from "helpers/command";
+import { installWrangler, retry, runCommand } from "helpers/command";
 import { debug } from "helpers/logging";
 import { detectPackageManager } from "helpers/packages";
 import { getProductionBranch, quoteShellArgs } from "./common";
-import type { TemplateConfig } from "./templateMap";
 import type { C3Context } from "types";
 
 /** How many times to retry the create project command before failing. */
@@ -21,21 +14,7 @@ const VERIFY_PROJECT_RETRIES = 3;
 
 const { npx } = detectPackageManager();
 
-export const runPagesGenerator = async (ctx: C3Context) => {
-	const frameworkConfig = ctx.template as unknown as TemplateConfig;
-
-	// Generate
-	const { configure } = frameworkConfig;
-
-	// Rectify discrepancies between installed node_modules and package specific
-	// lockfile before potentially adding new packages in `configure`
-	await resetPackageManager(ctx);
-
-	if (configure) {
-		await configure({ ...ctx });
-	}
-	chdir(ctx.project.path);
-
+export const runPagesGenerator = async () => {
 	// Install wrangler so that the dev/deploy commands work
 	await installWrangler();
 };
