+++ packages/create-cloudflare/src/workers.ts
@@ -1,28 +1,15 @@
-import { cp, mkdtemp, readdir, rm } from "fs/promises";
-import { tmpdir } from "os";
+import { existsSync, readdirSync } from "fs";
 import { join, resolve } from "path";
-import { processArgument } from "@cloudflare/cli/args";
 import { brandColor, dim } from "@cloudflare/cli/colors";
 import { spinner } from "@cloudflare/cli/interactive";
-import {
-	getWorkerdCompatibilityDate,
-	installWorkersTypes,
-	npmInstall,
-	installPackages,
-	runCommand,
-} from "helpers/command";
+import { getWorkerdCompatibilityDate, installPackages } from "helpers/command";
 import { appendFile, readFile, usesTypescript, writeFile } from "helpers/files";
 import { detectPackageManager } from "helpers/packages";
-import { chooseAccount } from "./common";
 import type { C3Context } from "types";
 
-const { dlx } = detectPackageManager();
+const { npm } = detectPackageManager();
 
 export const runWorkersGenerator = async (ctx: C3Context) => {
-	await npmInstall();
-	if (ctx.args.ts) {
-		await installWorkersTypes(ctx);
-	}
 	await installWorkersTypes(ctx);
 };
 
@@ -47,66 +34,6 @@ export async function appendToWranglerToml(ctx: C3Context, content: string) {
 
 	const wranglerTomlPath = resolve(ctx.project.path, "wrangler.toml");
 	appendFile(wranglerTomlPath, content);
-}
-
-async function copyExistingWorkerFiles(ctx: C3Context) {
-	await chooseAccount(ctx);
-
-	if (ctx.args.existingScript === undefined) {
-		ctx.args.existingScript = await processArgument<string>(
-			ctx.args,
-			"existingScript",
-			{
-				type: "text",
-				question:
-					"Please specify the name of the existing worker in this account?",
-				label: "worker",
-				defaultValue: ctx.project.name,
-			}
-		);
-	}
-
-	// `wrangler init --from-dash` bails if you opt-out of creating a package.json
-	// so run it (with -y) in a tempdir and copy the src files after
-	const tempdir = await mkdtemp(join(tmpdir(), "c3-wrangler-init--from-dash-"));
-	await runCommand(
-		[
-			...dlx,
-			"wrangler@3",
-			"init",
-			"--from-dash",
-			ctx.args.existingScript,
-			"-y",
-			"--no-delegate-c3",
-		],
-		{
-			silent: true,
-			cwd: tempdir, // use a tempdir because we don't want all the files
-			env: { CLOUDFLARE_ACCOUNT_ID: ctx.account?.id },
-			startText: "Downloading existing worker files",
-			doneText: `${brandColor("downloaded")} ${dim(
-				`existing "${ctx.args.existingScript}" worker files`
-			)}`,
-		}
-	);
-
-	// remove any src/* files from the template
-	for (const filename of await readdir(join(ctx.project.path, "src"))) {
-		await rm(join(ctx.project.path, "src", filename));
-	}
-
-	// copy src/* files from the downloaded worker
-	await cp(
-		join(tempdir, ctx.args.existingScript, "src"),
-		join(ctx.project.path, "src"),
-		{ recursive: true }
-	);
-
-	// copy wrangler.toml from the downloaded worker
-	await cp(
-		join(tempdir, ctx.args.existingScript, "wrangler.toml"),
-		join(ctx.project.path, "wrangler.toml")
-	);
 }
 
 async function installWorkersTypes(ctx: C3Context) {
