--- packages/create-cloudflare/templates/pre-existing/c3.json
@@ -1,16 +1,0 @@
-{
-	"configVersion": 1,
-	"id": "pre-existing",
-	"displayName": "Pre-existing Worker (from Dashboard)",
-	"platform": "workers",
-	"hidden": true,
-	"languages": ["js", "ts"],
-	"copyFiles": {
-		"js": {
-			"path": "../hello-world/js"
-		},
-		"ts": {
-			"path": "../hello-world/ts"
-		}
-	}
-}
