+++ packages/create-cloudflare/templates/pre-existing/c3.ts
@@ -1,0 +1,93 @@
+import type { C3Context } from "types";
+import { chooseAccount } from "../../src/common";
+import { processArgument } from "helpers/args";
+import { cp, mkdtemp, readdir, rm } from "fs/promises";
+import { join } from "path";
+import { tmpdir } from "os";
+import { runCommand } from "helpers/command";
+import { detectPackageManager } from "helpers/packages";
+import { brandColor, dim } from "@cloudflare/cli/colors";
+
+const config = {
+	configVersion: 1,
+	id: "pre-existing",
+	displayName: "Pre-existing Worker (from Dashboard)",
+	platform: "workers",
+	hidden: true,
+	languages: ["js", "ts"],
+	copyFiles: {
+		js: {
+			path: "../hello-world/js",
+		},
+		ts: {
+			path: "../hello-world/ts",
+		},
+	},
+	configure: async (ctx: C3Context) => {
+		await copyExistingWorkerFiles(ctx);
+	},
+};
+
+async function copyExistingWorkerFiles(ctx: C3Context) {
+	const { dlx, npm } = detectPackageManager();
+
+	await chooseAccount(ctx);
+
+	if (ctx.args.existingScript === undefined) {
+		ctx.args.existingScript = await processArgument<string>(
+			ctx.args,
+			"existingScript",
+			{
+				type: "text",
+				question:
+					"Please specify the name of the existing worker in this account?",
+				label: "worker",
+				defaultValue: ctx.project.name,
+			}
+		);
+	}
+
+	// `wrangler init --from-dash` bails if you opt-out of creating a package.json
+	// so run it (with -y) in a tempdir and copy the src files after
+	const tempdir = await mkdtemp(join(tmpdir(), "c3-wrangler-init--from-dash-"));
+	await runCommand(
+		[
+			...dlx,
+			"wrangler@3",
+			"init",
+			"--from-dash",
+			ctx.args.existingScript,
+			"-y",
+			"--no-delegate-c3",
+		],
+		{
+			silent: true,
+			cwd: tempdir, // use a tempdir because we don't want all the files
+			env: { CLOUDFLARE_ACCOUNT_ID: ctx.account?.id },
+			startText: "Downloading existing worker files",
+			doneText: `${brandColor("downloaded")} ${dim(
+				`existing "${ctx.args.existingScript}" worker files`
+			)}`,
+		}
+	);
+
+	// remove any src/* files from the template
+	for (const filename of await readdir(join(ctx.project.path, "src"))) {
+		await rm(join(ctx.project.path, "src", filename));
+	}
+
+	// copy src/* files from the downloaded worker
+	await cp(
+		join(tempdir, ctx.args.existingScript, "src"),
+		join(ctx.project.path, "src"),
+		{ recursive: true }
+	);
+
+	// copy wrangler.toml from the downloaded worker
+	await cp(
+		join(tempdir, ctx.args.existingScript, "wrangler.toml"),
+		join(ctx.project.path, "wrangler.toml")
+	);
+}
+
+export default config;
